+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    * Native support for `SMTP`, `Sendmail` and `Native` engines
    * Added ability to have 3rd party plugins that provide new engines (e.g. `SendGrid`, `MailJet`, etc.) with `SMTP`, `API` or `HTTPS` transports for faster and more reliable email sending
    * More flexible options for email formatting including RFC822 `name-addr` format (`Joe Bloggs <user@example.com>`) and simple arrays (`[user@example.com, Joe Bloggs]`)
+   * Support `tags` and `metadata` in `Email::buildMessage()`
 
 # v3.2.0
 ## 03/28/2022